Testing MySQL connection and inheritance in Python
This probably isn't the way Python wants things to work and it felt very
hacky getting this up and running.

At the minute I've added a "ConnectionManager" which probably should
just be called "ConfigurationManager" or some form of IO manager in
general. MySQL is a subclass which inherits from this (although
inheritance is a different beast in Python altogether)

Added additional requirements and updated the scrubber to test a MySQL
connection.